Two add image routing feature
There are now two image routes - one for the guest, and one for the user.
This is so that if the user has Authentication headers sent in, they will
be recognized as an Auth::user(), but if they do not have authentication
headers, they can upload on the public image uploading route that
automatically sets the user id to 0.