Database: Seeding (testing), Faker, and testing
Setup seeders that are automatically called by our DatabaseSeeder class to
automatically populate our images and users tables with test data that can
be used for testing purposes as well as local environment simulation of a
real world environment. Faker PHP library was installed so that we can
easily generate fake data for our application on such instances. Because
we do not know the IDs generated when Faker randomly creates images for
us, we changed ImageApiE2E's existingImages to now just pull random images
straight from the database instead of manual inputting of an existing
image's id in an array, allowing our testing environment to be reactive.