Controller/Model (User): OnDemand Hydration, reporting changes
Before, by default, the User Model, extending Ardent, automatically
hydrates inputs as our model when creating a new model. However, this was
posing a bunch of problems with Users when we were working with custom
fields, so rather, we just use the hydration when we need to by manually
editing the forceEntityHydrationFromInput property of the User model.
Also, we change the rules password for parts where we want to edit the
User Model, because for some reason, the default specified validation
spits out a "between" error when updating the model.